requirements.txt
numpy
biopython
